refactor: split sdl_render into sdl_texture module
